Refact: Offer-page
