+++ src/hooks/index.ts
@@ -7,6 +7,7 @@ export { useDocumentTitle } from './use-document-title/use-document-title';
 export { useMap } from './use-map/use-map';
 export { useCities } from './use-cities/use-cities';
 export { useOffers } from './use-offers/use-offers';
+export { useFetchOfferData } from './use-fetch-offer/use-fetch-offer';
 
 /** обертка не даст вызвать action которого нет у reducer, а так же для работы с асинхронными действиями */
 export const useAppDispatch = useDispatch<AppDispatch>;
